+++ testfarbenyay/src/main.rs
@@ -2,7 +2,6 @@ use farben::prelude::*;
 farben::load_styles!();
 
 fn main() {
-    cprintb!("[bold red]");
-    cprintln!("Bold red, [/red]Just bold.");
-    cprintln!("Shouldn't be bold red.");
+    init_styles();
+    cprintln!("[error] Bold red");
 }